StringOfItemCanBeConstructedFromString test
